Create a hook to get topics with prompts
